Added CascadingAuthenticationState to App.razor.
This fixes the following console error:

System.InvalidOperationException: Authorization requires a cascading parameter of type Task<AuthenticationState>. Consider using CascadingAuthenticationState to supply this.

Without it, users that are not logged in and hit a 404 will find that the client has crashed and it will not be redirected to the login page.